extract normalization tests
